Remove binaries
Keeping the binaries in the repo was intended for convenience but is
not scalable to an increasing number of language bindings and changes.

CI builds the binaries on every push to devel so in the future we could
provide a script to pull them on demand.

If there is a strong desire to include binaries in the repo despite
this, they can be added separately to the language-specific sub-repos.